First draft for user routes
