+++ node/package.json
@@ -18,7 +18,7 @@
     "@types/node": "^12.0.0",
     "@types/ramda": "^0.26.8",
     "@types/route-parser": "^0.1.2",
-    "@vtex/api": "6.28.1",
+    "@vtex/api": "6.28.2",
     "gocommerce.sitemap-app": "http://gocommerce.vtexassets.com/_v/public/typings/v1/gocommerce.sitemap-app@1.2.0/public/_types/react",
     "tslint": "^5.11.0",
     "tslint-config-vtex": "^2.1.0",
